Initial version of Create-BackwardCompatibleCaseInsensitiveHashtable.ps1
Uploading initial version of Create-BackwardCompatibleCaseInsensitiveHashtable.ps1, which allows the creation of hashtables (key-value pairs) that function just like those on newer versions of PowerShell - yet the approach taken in this function works on PowerShell v1, v2, and carries forward to the newest versions of PowerShell. Run this code if you need to use a hashtable and have your code work on any version of PowerShell.